+++ backend/backend/views.py
@@ -1,6 +1,4 @@
 from django.conf import settings
-from django.core.exceptions import ObjectDoesNotExist
-from django.db import transaction
 from django.urls import reverse
 from rest_framework import status
 from rest_framework.authtoken.views import ObtainAuthToken as DRFObtainAuthToken
@@ -16,8 +14,11 @@ from substrapp.orchestrator import get_orchestrator_client
 from substrapp.utils import get_owner
 from users.models.token import BearerToken
 from users.models.token import ImplicitBearerToken
+from users.models.token import get_implicit_bearer_token
 from users.serializers.token import BearerTokenSerializer
 from users.serializers.token import ImplicitBearerTokenSerializer
+
+from .exceptions import ImplicitLoginDisabled
 
 
 class ObtainBearerToken(DRFObtainAuthToken):
@@ -30,15 +31,15 @@ class ObtainBearerToken(DRFObtainAuthToken):
     throttle_classes = [AnonRateThrottle, UserLoginThrottle]
 
     def post(self, request, *args, **kwargs):
+        if not settings.EXPIRY_TOKEN_ENABLED:
+            raise ImplicitLoginDisabled()
+
         serializer = self.serializer_class(data=request.data, context={"request": request})
         serializer.is_valid(raise_exception=True)
         user = serializer.validated_data["user"]
 
-        with transaction.atomic():
-            token, just_created = ImplicitBearerToken.objects.get_or_create(user=user)
-            if not just_created:
-                token = token.handle_expiration()
-            return ApiResponse(ImplicitBearerTokenSerializer(token, include_payload=True).data)
+        token = get_implicit_bearer_token(user)
+        return ApiResponse(ImplicitBearerTokenSerializer(token, include_payload=True).data)
 
 
 class AuthenticatedBearerToken(DRFObtainAuthToken):
@@ -67,26 +68,23 @@ class ActiveBearerTokens(APIView):
             BearerTokenSerializer(token).data
             for token in BearerToken.objects.filter(user=request.user).order_by("-created")
         ]
-        try:
-            implicit_token = ImplicitBearerTokenSerializer(ImplicitBearerToken.objects.get(user=request.user)).data
-
-        except ObjectDoesNotExist:
-            implicit_token = None
+        implicit_tokens = [
+            ImplicitBearerTokenSerializer(token).data
+            for token in ImplicitBearerToken.objects.filter(user=request.user).order_by("-created")
+        ]
         return ApiResponse(
             {
                 "tokens": tokens,
-                "implicit_token": implicit_token,
+                "implicit_tokens": implicit_tokens,
             }
         )
 
     def delete(self, request, *args, **kwargs):
-        try:
-            token = BearerToken.objects.get(id=request.GET.get("id"))
-            if request.user == token.user:
-                token.delete()
+        for model in [BearerToken, ImplicitBearerToken]:
+            tokens = model.objects.filter(id=request.GET.get("id"))
+            if len(tokens) == 1 and request.user == tokens[0].user:
+                tokens[0].delete()
                 return ApiResponse(data={"detail": "Token removed"}, status=status.HTTP_200_OK)
-        except BearerToken.ObjectDoesNotExist or BearerToken.MultipleObjectsReturned:
-            pass
         return ApiResponse(data={"detail": "Token not found"}, status=status.HTTP_404_NOT_FOUND)