+++ tools/build_settings_doc.py
@@ -207,6 +207,7 @@ if __name__ == "__main__":
     args = parse_arguments()
     settings = {}
     settings["Global"] = load_settings_from_file(SETTINGS_FOLDER / "common.py")
+    settings["Global"] += load_settings_from_file(SETTINGS_FOLDER / "deps/restframework.py")
     settings["JWT"] = load_settings_from_file(SETTINGS_FOLDER / "deps/jwt.py")
     settings["Orchestrator"] = load_settings_from_file(SETTINGS_FOLDER / "deps/orchestrator.py")
     settings["Task broker"] = load_settings_from_file(SETTINGS_FOLDER / "deps/celery.py")